second commit

+++ backend/routes.php
@@ -1,7 +1,6 @@
 <?php
 
-class Router
-{
+class Router {
     private $basePath;
     private $requestMethod;
     private $requestUri;
@@ -9,8 +8,7 @@ class Router
     private $pathParts;
 
     // Инициализирует переменные и подготавливает путь для обработки
-    public function __construct($basePath = '/')
-    {
+    public function __construct($basePath = '/') {
         $this->basePath = $basePath;
         $this->requestMethod = $_SERVER["REQUEST_METHOD"];
         $this->requestUri = $_SERVER["REQUEST_URI"];
@@ -22,13 +20,13 @@ class Router
     }
 
     // Основная функция, которая решает, какой маршрут обработать
-    public function handleRequest()
-    {
-        // error_log("PATH: " . $this->path);
-
+    public function handleRequest() {
         if ($this->pathParts[0] === 'api' && $this->pathParts[1] === 'ads') {
             // Обрабатывает API-запросы
             $this->handleApiRequest();
+        } elseif ($this->path === 'bot') {
+            // Обработка вебхуков для Telegram
+            $this->handleTelegramWebhook();
         } elseif ($this->requestMethod === 'GET' && empty($this->path)) {
             // Отправляет index.html при запросе на корневой путь
             header('Content-Type: text/html');
@@ -40,59 +38,48 @@ class Router
     }
 
     // Обрабатывает API-запросы
-    private function handleApiRequest()
-    {
+    private function handleApiRequest() {
         $controller = new AdController();
 
         if (count($this->pathParts) === 2) {
-            // Обработка запросов к коллекции объявлений или создание нового объявления
-            $this->handleAdsCollectionOrCreateRequest($controller);
+            // Обработка коллекции объявлений или создание нового объявления
+            if ($this->requestMethod === 'GET') {
+                $controller->getAllAds();
+            } elseif ($this->requestMethod === 'POST') {
+                $controller->createAd();
+            } else {
+                http_response_code(405);
+                echo json_encode(["message" => "Method Not Allowed"]);
+            }
         } elseif (count($this->pathParts) === 3 && is_numeric($this->pathParts[2])) {
-            // Обработка запросов для отдельного объявления
-            $this->handleSingleAdRequest($controller);
+            // Обработка конкретного объявления по ID
+            $adId = intval($this->pathParts[2]);
+            if ($this->requestMethod === 'GET') {
+                $controller->getAd($adId);
+            } elseif ($this->requestMethod === 'DELETE') {
+                $controller->deleteAd($adId);
+            } else {
+                http_response_code(405);
+                echo json_encode(["message" => "Method Not Allowed"]);
+            }
+        } elseif (count($this->pathParts) === 3 && $this->pathParts[2] === 'hash_num'
+            && $this->requestMethod === 'POST') {
+            // Обработка запросов к коллекции объявлений по пользователю (по hash_num)
+            $controller->findAdsByUser();
         } else {
             http_response_code(404);
             echo json_encode(["message" => "Not Found"]);
         }
     }
 
-    // Обрабатывает запросы к коллекции объявлений (GET) или создание нового объявления (POST)
-    private function handleAdsCollectionOrCreateRequest($controller)
-    {
-        switch ($this->requestMethod) {
-            case 'GET':
-                // Получить все объявления с параметрами пагинации
-                $controller->getAllAds();
-                break;
-            case 'POST':
-                // Создать новое объявление
-                $controller->createAd();
-                break;
-            default:
-                http_response_code(405);
-                echo json_encode(["message" => "Method Not Allowed"]);
-                break;
-        }
+    // Обрабатывает вебхуки для Telegram
+    private function handleTelegramWebhook() {
+        $controller = new TgController();
+
+        // Метод для обработки вебхуков
+        $controller->handleWebhook();
     }
 
-    // Обрабатывает запросы к конкретному объявлению (GET, DELETE)
-    private function handleSingleAdRequest($controller)
-    {
-        $adId = intval($this->pathParts[2]);
-
-        switch ($this->requestMethod) {
-            case 'GET':
-                $controller->getAd($adId);
-                break;
-            case 'DELETE':
-                $controller->deleteAd($adId);
-                break;
-            default:
-                http_response_code(405);
-                echo json_encode(["message" => "Method Not Allowed"]);
-                break;
-        }
-    }
 }
 
 // Создание и запуск роутера
